refactor: extract Safe wallet confirmation into safeWalletUtils
Move fibonacciDelays, abortableSleep, pollForContractEvent,
ResilientConfirmOptions, and resilientConfirm out of rpcUtils.ts into
a dedicated safeWalletUtils.ts. rpcUtils.ts now only contains RPC
transport (raceTransport, raceViemProviderBuilder, withWcRpcFirst),
gas estimation (preEstimateGasForEvmTxs), and chain switching
(waitForChainSwitch, ensureWalletOnChain).

Corresponding tests moved to safeWalletUtils.test.ts.
useTokenTransfer.ts imports resilientConfirm from the new module.
