+++ src/features/transfer/__tests__/useTokenTransfer.test.ts
@@ -151,7 +151,10 @@ vi.mock('../../chains/utils', () => ({
 vi.mock('../../chains/rpcUtils', () => ({
   ensureWalletOnChain: vi.fn(() => Promise.resolve()),
   preEstimateGasForEvmTxs: vi.fn(),
-  // Pass through to wallet confirm — resilientConfirm is tested in rpcUtils.test.ts
+}));
+
+vi.mock('../../chains/safeWalletUtils', () => ({
+  // Pass through to wallet confirm — resilientConfirm is tested in safeWalletUtils.test.ts
   resilientConfirm: vi.fn((walletConfirm: () => Promise<any>) => walletConfirm()),
 }));
